requirements.txt
requests
streamlit
google-genai
python-dotenv